Create module for Key and for ALPHABET related actions
